Move Crosslink to attestation types file
