+++ beaconChain/tests/types/consistency.test.ts
@@ -60,7 +60,6 @@ describe("types", () => {
   const vars = {};
   for (const name in types) {
     vars[name] = types[name];
-    if(name === "latestAttestations") console.log(`TYPE: ${types[name]}`)
   }
   // Now that we have an object of interfaces and and object of runtime type variables, we can perform our tests
   it("Every interface should have a corresponding runtime type variable", () => {
@@ -96,10 +95,6 @@ describe("types", () => {
     Object.values(interfaces)
       .map((iface: any) => {
         const rtVar = vars[iface.name];
-        if(iface.name === 'BeaconState') {
-            console.log(JSON.stringify(rtVar))
-            console.log(JSON.stringify(iface))
-        }
 
         assert(
           rtVar.fields.length === iface.fields.length,
