+++ beaconChain/types/state.ts
@@ -13,6 +13,7 @@ import {
 } from "./primitive";
 
 import {
+  Crosslink,
   PendingAttestation,
 } from "./attestation";
 
@@ -64,20 +65,6 @@ export const Validator = {
     ["withdrawalEpoch", uint64],
     ["slashedEpoch", uint64],
     ["statusFlags", uint64],
-  ],
-};
-
-export interface Crosslink {
-  // Slot number
-  epoch: uint64;
-  // Shard chain block hash
-  shardBlockRoot: bytes32;
-}
-export const Crosslink = {
-  name: "Crosslink",
-  fields: [
-    ["epoch", uint64],
-    ["shardBlockRoot", bytes32],
   ],
 };
